Rename C++ namespace
