+++ src/amulet/pybind11/compatibility.hpp
@@ -10,7 +10,8 @@ static std::string get_version_top(std::string version)
     return version.substr(0, version.find('.', version.find('.') + 1));
 }
 
-namespace pybind11_extensions {
+namespace Amulet { {
+namespace pybind11 {
 
 void init_compiler_config(pybind11::module m)
 {
@@ -50,4 +51,5 @@ void check_compatibility(pybind11::module a, pybind11::module b)
     }
 }
 
-}
+} // namespace pybind11
+} // namespace Amulet
