+++ src/amulet/pybind11/types.hpp
@@ -6,8 +6,10 @@
 // This is used as a return in comparison operators.
 
 
-namespace pybind11_extensions {
+namespace Amulet { {
+namespace pybind11 {
     namespace types {
         using NotImplementedType = pybind11_extensions::PyObjectStr<"types.NotImplementedType">;
     }
-}
+} // namespace pybind11
+} // namespace Amulet
